Extract <Skill /> to own component
